Lots of fixes from use in the last week
Main thing was to solve the problem of missed intervals due to commands
being lost and to implement a less naive control loop and separate
control and UI update loops.

The trainer would sometimes not register target power commands,
despite the acknowledged message tx completing successfully. Instead
of "send and forget", implement a loop that queries the command status
and resends commands on failure timeout, until the status matches the
command.